fix error naming and separate files
