+++ Sources/Service/MusicLibraryService.swift
@@ -1,111 +1,7 @@
 import MediaPlayer
 
-/// Protocol for querying media items from the music library.
-///
-/// This protocol defines the service layer interface for fetching media items with flexible filtering and grouping.
-/// It is primarily used by `MusicLibrary` for the concrete implementation, but can be injected for testing or custom implementations.
-///
-/// ## Usage
-/// The protocol is designed for dependency injection:
-/// - Use ``MusicLibraryServiceProtocol/live`` in production
-/// - Pass mock implementations in tests
-/// - Implement your own conforming type for custom behavior
-///
-/// ## Generic Types
-///
-/// The protocol defines two associated types for flexibility:
-///
-/// **Error Type (`E`):** Allows concrete implementations to define their own error types,
-/// enabling type-safe error handling specific to each implementation.
-///
-/// **Query Type (`Q`):** Allows implementations to use different query types conforming to
-/// `MediaQueryProtocol`. This enables using `MPMediaQuery` in production and mock queries in tests.
-///
-/// ## Example
-///
-/// Using with a mock query for testing:
-/// ```swift
-/// let service: any MusicLibraryServiceProtocol = MusicLibraryService<MockMediaQueryWithTwoItems>()
-/// let songs = try await service.fetchAll(.music, groupingType: .album)
-/// ```
-public protocol MusicLibraryServiceProtocol: Sendable {
-    associatedtype E: Error
-    associatedtype Q: MediaQueryProtocol
-
-    /// Fetches all media items of a specific type with optional grouping.
-    ///
-    /// Retrieves all media items matching the specified media type from the music library.
-    /// Results can be grouped by title, album, artist, or other criteria.
-    ///
-    /// - Parameters:
-    ///   - type: The type of media to fetch (typically `.music`)
-    ///   - groupingType: How to organize the returned items
-    /// - Returns: Array of media items matching the type
-    /// - Throws: An error of specific type if the query fails 
-    ///
-    /// ## Example
-    /// ```swift
-    /// let items = try await service.fetchAll(.music, groupingType: .album)
-    /// ```
-    func fetchAll(
-        _ type: MPMediaType,
-        groupingType: MPMediaGrouping
-    ) async throws -> [MPMediaItem]
-
-    /// Fetches media items collection matching a predicate with optional filtering.
-    ///
-    /// - Parameters:
-    ///   - type: The type of media to fetch (typically `.music`)
-    ///   - predicate: The predicate to filter items
-    ///   - comparisonType: How to compare the predicate value (e.g., `.equalTo`, `.contains`)
-    ///   - groupingType: How to organize the returned items
-    /// - Returns: Array of media items matching the predicate
-    /// - Throws: An error of specific type if the query fails
-    func fetchCollections(
-        _ type: MPMediaType,
-        with predicate: MediaItemPredicateInfo,
-        comparisonType: MPMediaPredicateComparison,
-        groupingType: MPMediaGrouping
-    ) async throws -> [MPMediaItemCollection]
-
-    /// Fetches all media collections of a specific type with optional grouping.
-    func fetchAllCollections(
-        _ type: MPMediaType,
-        groupingType: MPMediaGrouping
-    ) async throws -> [MPMediaItemCollection]
-
-    /// Fetches media items matching a predicate with optional filtering.
-    ///
-    /// Queries the music library for items that match the provided predicate and media type.
-    /// The comparison type determines how the predicate is matched.
-    ///
-    /// - Parameters:
-    ///   - type: The type of media to fetch (typically `.music`)
-    ///   - predicate: The predicate to filter items
-    ///   - comparisonType: How to compare the predicate value (e.g., `.equalTo`, `.contains`)
-    ///   - groupingType: How to organize the returned items
-    /// - Returns: Array of media items matching the predicate
-    /// - Throws: An error of specific type if the query fails
-    ///
-    /// ## Example
-    /// ```swift
-    /// let items = try await service.fetch(
-    ///     .music,
-    ///     with: .artist("Taylor Swift"),
-    ///     comparisonType: .contains,
-    ///     groupingType: .album
-    /// )
-    /// ```
-    func fetch(
-        _ type: MPMediaType,
-        with predicate: MediaItemPredicateInfo,
-        comparisonType: MPMediaPredicateComparison,
-        groupingType: MPMediaGrouping
-    ) async throws -> [MPMediaItem]
-}
-
 /// Factory extension for creating a live instance of `MusicLibraryService`.
-extension MusicLibraryServiceProtocol where Self == MusicLibraryService<MPMediaQuery>, E == Self.MusicLibraryServiceError {
+extension MusicLibraryServiceProtocol where Self == MusicLibraryService<MPMediaQuery>, E == MusicLibraryServiceError {
     /// Returns a live instance of the default `MusicLibraryService` implementation.
     ///
     /// Use this property in production code to get the standard service implementation:
@@ -133,39 +29,6 @@ public final class MusicLibraryService<T: MediaQueryProtocol>: MusicLibraryServi
     public typealias E = MusicLibraryServiceError
     public typealias Q = T
 
-    /// Errors that can occur during music library service operations.
-    public enum MusicLibraryServiceError: Error, LocalizedError, Equatable {
-        /// No songs were found matching the query.
-        case noSongsFound
-
-        /// No collections were found matching the query.
-        case noCollectionsFound
-
-        /// No album was found matching the specified predicate.
-        case noCollectionFound(MediaItemPredicateInfo)
-
-        /// No song was found matching the specified predicate.
-        case noSongFound(MediaItemPredicateInfo)
-
-        /// A human-readable error description suitable for displaying to users.
-        public var errorDescription: String? {
-            switch self {
-            case .noSongsFound:
-                "No songs found in your music library. Try adding music to your library and try again."
-            case .noCollectionsFound:
-                "No media collections found in your music library. Try adding music to your library and try again."
-            case .noCollectionFound(let predicate):
-                "Couldn't find an album matching \(predicate.description). Check your filters and try again."
-            case .noSongFound(let predicate):
-                "Couldn't find a song matching \(predicate.description). Check your filters and try again."
-            }
-        }
-
-        public static func == (lhs: MusicLibraryService<T>.MusicLibraryServiceError, rhs: MusicLibraryService<T>.MusicLibraryServiceError) -> Bool {
-            lhs.errorDescription == rhs.errorDescription
-        }
-    }
-
     // MARK: - MusicLibraryServiceProtocol Implementation
 
     /// Fetches media items matching a predicate with default parameters.
@@ -178,10 +41,10 @@ public final class MusicLibraryService<T: MediaQueryProtocol>: MusicLibraryServi
         comparisonType: MPMediaPredicateComparison = .equalTo,
         groupingType: MPMediaGrouping = .title
     ) async throws -> [MPMediaItem] {
-        guard let songs = query(type, withFilter: predicate, comparisonType, groupingType).items else {
-            throw E.noSongFound(predicate)
+        guard let items = query(type, withFilter: predicate, comparisonType, groupingType).items else {
+            throw E.noItemFound(predicate)
         }
-        return songs
+        return items
     }
 
     /// Fetches all media items of a specific type with grouping.
@@ -192,27 +55,30 @@ public final class MusicLibraryService<T: MediaQueryProtocol>: MusicLibraryServi
         _ type: MPMediaType,
         groupingType: MPMediaGrouping
     ) async throws -> [MPMediaItem] {
-        guard let songs = query(type, groupingType).items else {
-            throw E.noSongsFound
+        guard let items = query(type, groupingType).items else {
+            throw E.noItemsFound
         }
-        return songs
+        return items
     }
 
     /// Fetches all media collections of a specific type with grouping.
     ///
-    /// Implementation of ``MusicLibraryServiceProtocol/fetchAll(_:groupingType:)`` that retrieves all items of the specified type
+    /// Implementation of ``MusicLibraryServiceProtocol/fetchAllCollections(_:groupingType:)`` that retrieves all collections of the specified type
     /// without additional filtering, organized by the specified grouping type.
     public func fetchAllCollections(
         _ type: MPMediaType,
         groupingType: MPMediaGrouping
     ) async throws -> [MPMediaItemCollection] {
-        guard let songs = query(type, groupingType).collections else {
+        guard let collections = query(type, groupingType).collections else {
             throw E.noCollectionsFound
         }
-        return songs
+        return collections
     }
 
     /// Fetches media item collection matching a predicate with default parameters.
+    ///
+    /// Implementation of ``MusicLibraryServiceProtocol/fetchCollections(_:with:comparisonType:groupingType:)`` with default comparison type (`.equalTo`)
+    /// and default grouping type (`.title`). Combines the media type predicate with the user-provided predicate.
     public func fetchCollections(
         _ type: MPMediaType,
         with predicate: MediaItemPredicateInfo,
@@ -259,4 +125,3 @@ public final class MusicLibraryService<T: MediaQueryProtocol>: MusicLibraryServi
         return query
     }
 }
-
